+++ Editor/Package.cs
@@ -6,10 +6,11 @@ namespace Hananoki.BuildAssist {
   public static class Package {
     public const string name = "BuildAssist";
     public const string editorPrefName = "Hananoki.BuildAssist";
-    public const string version = "1.2.0";
+    public const string version = "1.3.0";
     public static string projectSettingsPath => $"{Environment.CurrentDirectory}/ProjectSettings/BuildAssist.json";
   }
   
+#if UNITY_EDITOR
   [EditorLocalizeClass]
   public class LocalizeEvent {
     [EditorLocalizeMethod]
@@ -22,4 +23,5 @@ namespace Hananoki.BuildAssist {
       }
     }
   }
+#endif
 }